+++ packages/node-core/src/api.service.ts
@@ -25,13 +25,14 @@ export interface IApiConnectionSpecific<A = any, SA = any, B extends Array<any> 
   handleError(error: Error): ApiConnectionError;
   apiConnect(): Promise<void>;
   apiDisconnect(): Promise<void>;
+  updateChainTypes?(chainTypes: unknown): Promise<void>;
 }
 
 export abstract class ApiService<
   A = any,
   SA = any,
   B extends Array<any> = any[],
-  Connection extends IApiConnectionSpecific<A, SA, B> = IApiConnectionSpecific<A, SA, B>
+  Connection extends IApiConnectionSpecific<A, SA, B> = IApiConnectionSpecific<A, SA, B>,
 > implements IApi<A, SA, B>
 {
   constructor(