+++ packages/node-core/src/indexer/connectionPool.service.ts
@@ -250,4 +250,11 @@ export class ConnectionPoolService<T extends IApiConnectionSpecific<any, any, an
     this.lastCacheFlushTime = Date.now();
     await this.handleConnectionStateChange();
   }
+
+  async updateChainTypes(newChainTypes: unknown): Promise<void> {
+    for (const endpoint in this.allApi) {
+      await this.allApi[endpoint].updateChainTypes?.(newChainTypes);
+    }
+    logger.info(`Network chain types updated!`);
+  }
 }